Add buffer cancellation API
The iio_buffer_refill() and iio_buffer_push() functions return when either
the specified amount of data has been received / transmitted or when a
timeout occurs. The timeout is usually used by applications to protect
themselves from the case where the data stream is blocked, e.g. malfunction
of the converter, so that the application does not block indefinitely. If
the data stream runs dry the timeout will eventually be triggered and the
application can recover.

This approach works fine for continuous data streams with a constant data
rate since there is a predictable time it will take to complete the
iio_buffer_refill() and iio_buffer_push() operations and the timeout can be
set accordingly. It does not work very well for bursty data streams or when
the start of the data stream is triggered by external stimuli not under
software control. In this case there often is no sensible timeout as it is
not predictable how long it will take the refill() or push() operation to
complete. If an application sets a high or no timeout it risks blocking,
non-responsiveness and deadlocking, if the application sets a low timeout
it risks missing valid data that would have been received after the timeout
expired.

To address this issue introduce the new iio_buffer_cancel() API. Calling
iio_buffer_cancel() will ensure that all currently blocking refill() and
push() operations on the same buffer will wakeup. All future invocations on
the same buffer will immediately return with an error. This avoids
potential ordering issues and an application that called iio_buffer_cancel()
can be sure that no other operations on the same buffer will block. A
obvious side effect of the cancel() operation is that it is no longer
possible to use the buffer for that transmission or receiving. An
application that wishes to start a new data stream for the same device has
to destroy the current buffer and create a new one.

Due to its nature the iio_buffer_cancel() operation has to be called from a
different thread than the one that executes the iio_buffer_push() or
iio_buffer_refill() operation. It is not safe to call the cancel()
operation from signal handler context.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>