local: Always enable buffer when it is created
Currently when using the high-speed interface the buffer is enabled when it
is created. When not using the high-speed interface the buffer is enabled
right before the first read or write. It's not exactly clear why the later
is done, except that it was required to support cyclic write over a
non-standard interface which was using file IO to setup the buffer.

Always enabling the buffer when it is created makes the behaviour between
the high-speed and non high-speed mode more consistent. While careful
analysis reveals no issues, lets hope this change does not break anything.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>