local: Abort early when using cyclic mode on non high-speed device
Currently when trying to use cyclic mode on a non high-speed device
creating the buffer will work fine, but when trying to write the data to
the device an error is returned.

Now that cyclic mode is only supported on high-speed devices abort early
and return an error when the buffer is created. This is more useful
information for diagnostic purposes. If the requested buffer type is not
supported by the device creating the buffer should fail.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>