local: Inline {read,write}_all() in local_{read,write}
local_{read,write} are the only callers of {read,write}_all() and neither
functions are two large that the flow is improved by keeping them separate.

In order to fix some issues it is necessary to interleave sections of both
functions. This will be easier to do when they are contained within a
single function.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>