local: Add cancellation support
Add support for the iio_buffer_cancel() API to the local backend. This is
implemented by using a eventfd() which is also poll()-ed on when waiting
for IO from the IIO device. When the buffer is canceled a event is written
to the eventfd but never read. This means any currently active or future IO
attempts will wake up immediately.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>