local: Fix corner cases of timeout handling
When using a buffer in blocking mode poll() is used for the timeout
handling. We wait until either the timeout occurs or data/room in buffer
becomes available. poll() will wakeup as soon as there is some data/room
available. It is possible though that this is not enough data/room for the
whole buffer. In this case another read()/write() call is issued until
enough data/room is available. The later step bypasses the timeout handling
and if some data/room becomes available, but not enough the operations will
block past the timeout.

To avoid this instead of directly going back to read()/write() recalculate
the remaining timeout and call poll() again. poll() will either wake up
when more data/room becomes available or the timeout happens.

To ensure that read()/write() never block and blocking is always done in
poll() open the file descriptor in non-blocking mode unconditionally.

Similar issues apply to the BLOCK_DEQUEUE ioctl, so it receives the same
fix.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>