iio_readdev: Add cancellation support
Add cancellation support to the iio_readdev tool. When user input is
received that the application should terminate issue a iio_buffer_cancel(),
this will make sure that the applications terminates immediately. Otherwise
it has to wait for the current iio_buffer_refill() operation to complete or
for a timeout to occur.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>