utilities: Make set_blocking_mode() non-available on Windows
On Windows platforms set_blocking_mode() always fails. It is also never
called on a code path that is built into the library when building it on
Windows.

Instead of letting the function return -ENOSYS simply don't make it
available on Windows. This will find incorrect usages at compile-time
rather than run-time.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>