network: Add cancellation support
Add cancellation support for the network backend. This is implemented by
adding an additional wakeup event source. When the buffer is cancelled the
event source is triggered.  Blocking recv() and send() operations are
replaced by waiting for either the cancel event to be triggered or the
socket to become ready. If the cancel event is triggered the IO operation
is aborted, otherwise if the socket becomes ready the recv() or send()
operation is performed non-blocking.

The specifics of how this is implemented are operating system depending. On
Windows the WSAEvent API is used. On Linux eventfd() and poll() are used if
eventfd() is available. On all other systems pipe() and poll() are used.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>