Local backend: Hide private attributes of scan-element channels
The scan-element channels feature a couple of attributes that are
exposed by the libiio API:
- "index" can be obtained using iio_channel_get_index,
- "type" can be obtained using iio_channel_get_data_format.

The special "en" attribute is not exposed by the libiio API, and should
not.

There is a problem with keeping those attributes in the iio_channel: it
is possible that a shared attribute will end up having the same short
name, causing libiio to misbehave. By hiding the scan-element
attributes, we make sure that this conflict will not happen.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>
Reported-by: Lucas Magasweran <lucas.magasweran@daqri.com>
Tested-by: Lucas Magasweran <lucas.magasweran@daqri.com>