Network backend: Use -EPIPE instead of -EAGAIN on timeout
The -EAGAIN error code would be used if we were doing non-blocking I/O,
which the network backend does not support.

Besides, before timeout detection was fixed (in 3b53ec6), -EPIPE was the
error code returned after a timeout event. This commit restores the old
behaviour to avoid breaking the API with older versions of libiio.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>