buffer: Fail early when trying to create a 0 samples buffer
Trying to create a 0 samples sized buffer will always fail. Currently the
error is only handled at the lowest levels of the stack when trying to
create the kernel buffer.

Instead detect this type of error early on in iio_device_create_buffer().
This means the individual backends do not have to care about the error and
we also get consistent behaviour across backends in terms of the error
code returned.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>