network: network_do_splice(): Fix read/write imbalance
The current code in network_do_splice() assumes that the write splice
operation will always write as much data as requested. This is not
necessarily true and if that happens 'len' will only be decremented by the
amount of actually written. But 'len' is also used for the read splice
operation which will now attempt to read more data than actually requested.
Which will cause the whole operation to fail.

To fix this track the amount of data to read and write independently.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>