iiod-client: When attribute data is too big to read, discard data
IIOD will send the length of the data to read then send the data
straight away without waiting for confirmation.

However, when the C code calls iio_device_attr_read (or
iio_channel_attr_read, or variants) with a buffer size inferior to the
actual data size, the backends using iiod-client would return an error
without actually discarding the data, which disrupted the communication.

Now, when an attribute value contains too much data to read, we just
discard that data so that the communication isn't affected.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>