IIOD: Introduce thread pools
The thread pool object can be used to manage a number of threads.

Zero, one or more threads can be added to the thread pool. Each thread
will be passed a pointer to the pool, from which it can obtain a
pollable file descriptor using thread_pool_get_poll_fd(), that will
notify the thread when it should exit.

To stop all threads in the pool, you can call thread_pool_stop(). Since
threads are passed a pointer to the pool object, they can request the
STOP event themselves.

The function thread_pool_stop_and_wait() is slightly different in that
it will request the STOP event and then wait for all the threads to
exit. Needless to say, this shouldn't be called on a thread pool that
contains the current thread or a deadlock will happen.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>