IIOD: Fix read_line() function
The read_line() function is optimized so that it can read a whole line
of text in just two system calls, instead of reading
character-by-character until the end-of-line character is found.

However, it did not handle the fact that recv() can return a number of
bytes smaller than the actual command being sent. In that case it was
mis-detecting the beginning of the command line as garbage, since the \n
was not found.

This commit makes sure that the read_line() function will call again
recv() if the \n is not found and we still have space in the destination
buffer.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>