iiod: {read,write}fd_io(): Fix loop continuation
In the very unlikely event poll() woke up, but none of the fds are ready we
want to go back to the beginning of the loop and start another poll().
This is done using the continue statement. This will trigger a evaluation
of the loop condition though and at this point both ret and errno contain
bogus values which will cause undefined behaviour.

To avoid this slightly refactor the way the loop condition is checked by
moving it outside of the while () statement and checking it directly before
manually. This will make sure that they are not checked when the continue
statement is used.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>