iiod: Protect against aio race condition
async_io() is not safe to be called from multiple threads concurrently for
the same AIO context. When being executed concurrently it is possible that
one thread steals the completion event of another thread.  This other
thread will no wait forever for a completion event that will never be
delivered.

Usually this is not an issue as IO operations on the same context are
synchronized to each other. There is one exception though when a remote
buffer transmit/receive is cancelled all threads waiting for IO on that
connection are woken up. Under these circumstances it is possible that both
the rw_thd() and the rw_buffer() function try to access async_io() at the
same time. Both calls will eventually fail since the channel has been
closed, but not providing mutual exclusion can result in the deadlock
described above.

To avoid this add mutex that prevents multiple threads from scheduling a
async IO operation on the same AIO context at the same time.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>