iiod: Run rw_thd thread in main thread pool
The rw thread is shared between all clients which are transferring data
from or to the same device. It is created when the first client connects
and will stay active until all clients using it have exited.

Currently the rw thread is created in the thread pool of the first client
that connects. If a second client belonging to a different thread pool
connects it will attach to the existing rw thread. Now if the first client
exits the rw thread will stay active, which means running
thread_pool_stop_and_wait() for the thread pool for this client will block
until the second client exits as well. This can block resources (e.g. like
a USB pipe).

To avoid this launch the rw thread in the main thread pool, so only exiting
the application has to wait for the rw threads to exit, but not individual
clients.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>