iio-monitor: Set locale before parsing double values from sysfs
strtod() uses the current locale when parsing numbers, while all values
in sysfs use the "C" locale.  Hence the former may expect a comma as the
radix character, while the latter always uses a decimal point, leading
to the display of incorrect values.

Temporarily switch to the "C" locale before parsing double numbers to
fix this.

Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>