Use macro iio_snprintf instead of snprintf
The problem of using snprintf directly is that Visual Studio will
complain that we're not using the more secure sprintf_s variant.

The way to work around that has been to redefine 'snprintf' as a macro
set to 'sprintf_s', so that all calls in the code to snprintf would
resolve as calls to the sprintf_s macro.

However, Visual Studio 2015 complains that we're shadowing a standard
library function. The solution here, is to change all calls to snprintf
in the code to a macro iio_snprintf that we can point either to
sprintf_s if available, or snprintf.

Fixes #89.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>