network: Implement connect timeout for Windows
Currently a connect timeout in the network backend is only implemented for
non-Windows systems. This patch implements the connect timeout for Windows
ensuring consistent behavior across different platforms.

It is possible to use the same general approach as on non-Windows systems,
except that the method for making a socket non-blocking as well as the
return code for a in-progress connect() operation are different on
Windows.

These differences are hidden by implementing them in architecture specific
helper functions.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>