network: Use default timeout during buffer creation
Creating a new buffer should normally not block beyond normal network
communication latencies. But if for whatever reason the connection becomes
blocking (e.g. network cable unplugged) there is no way to recover and the
buffer creation will block for the specified timeout. This might be
indefinite in which case the application can't recover since it is not
possible to cancel buffer creation.

Considering this establish the new connection for the buffer with the
default timeout and only switch to the configured timeout once the buffer
has been successfully opened and it is clear that the buffer creation will
succeed.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>