Use channel number instead of channel index in mask
By 'channel number', I mean the index at which the 'iio_channel'
structure appears in the parent iio_device.

Most of the cases, this is generally the same as the index of the
channel (returned by iio_channel_get_index()), since the channels are
ordered by index in the parent iio_device.

This change is required to support the (rare) cases where the
iio_channels share the same index. Previously, those channels would
set the exact same bit in the channel mask, which would confuse IIOD
about which channels were enabled.

This change should not break the ABI between libiio and IIOD, because
the channels are sorted the exact same way on both ends.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>