USB backend: Support URB transfer of 1 MiB max
If the size of the data to transfer is too big, the
IOCTL_USBFS_SUBMITURB ioctl (called by libusb) might fail with
errno set to ENOMEM, as the kernel might use contiguous allocation
for the URB if the driver doesn't support scatter-gather.
To prevent that, we support URBs of 1 MiB maximum. The iiod-client
code will handle this properly and ask for a new transfer.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>