Local backend: Return -EACCES if we don't have permission to access attr
This might break some setups.

Before, running an application as root would have granted it rights to
read / write all the attributes independently of the permissions set to
their corresponding sysfs files, even if the system call would have
failed, generally with an error code like -ENODEV or -ENOENT.

Now, trying to access the attributes while running an application as a
user, including root, that does not have the necesary rights for it will
return a -EACCES error code.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>