local: Don't clear channel enabled state on buffer destruction
Commit 52736b905624 ("Local backend: Disable individual channels when
closing devices") changed the behaviour of the local backend so that all
channels are marked as disabled when a buffer is destroyed (note that the
commit message says that they are disabled when the device is closed).

This breaks code that repeatedly creates and destroys buffer without
changing the channel enabled configuration. The same code work fine on
non-local backends since there the channel state is not cleared when the
buffer is destroyed.

Restore the correct and expected behavior in the local backend.

The system global channel enabled state is still cleared when the buffer is
destroyed, but the context local state is preserved and will be re-applied
to the global state once the next buffer is created.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>