RFC: Fix channel identifier heuristics
libiio currently has issues identifying the channel identifier (channel_id)
when a driver reports multiple attributes of the same type (e.g.
IIO_CONCENTRATION) with a name extension (.extend_name field in
struct iio_chan_spec) as it tries to resolve the name extension as the
modifier (e.g. IIO_MOD_VOC). In this case the startup routine
(local_create_context) fails to establish a valid context and aborts.

This patch defuses the situation by resolving the second-to-last underscore
(_) separated segment as the modifier, leaving room for a name extension
that itself potentially contains underscores.

* Modify find_channel_modifier to find the modifier in the second-to-last
  underscore separated segment.
  Also, report the offset from the beginning of the string until the
  position of the modifier, or the expected position of the modifier in
  case of IIO_NO_MOD.
* In is_global_attr, check the string prefix until the last underscore for
  a match