local: Return error when no blocks could be allocated
The BLOCK_ALLOC_IOCTL will not fail if not enough memory is available to
allocate all requested memory. Instead it will allocate as much as it can
and report how much was allocated.

This includes the case where no memory could be allocated. A buffer without
any backing memory doesn't make much sense though, so error out in this
case.

This also fixes a deadlock in libiio where it would be waiting for a block
to become available during buffer creation (which would never happen
without any blocks).

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>