build: generate static library for libiio
For some environments/systems, it's preferred to have a static
libiio library to link into the app.
Especially when the version of libiio installed on
the system/distro is older than the one that's required.

This change enables the generation of a static lib,
along side the shared lib.
Users can choose to use it [or not].

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>