network: Fix buffer reuse
iiod_client_open_unlocked() is executed before setup_cancel() is called.
This means the cancellable flag needs to be set to false while running
iiod_client_open_unlocked() otherwise undefined behavior might occur.

The cancellable flag is a per device flag that at the moment is only set to
false then the IIO network context is created and is set to true when a
buffer as been successfully create. This means when creating a buffer for a
device, destroying the buffer and then creating another buffer for the same
device the cancellable flag will be set to true when running
iiod_client_open_unlocked().

To avoid this set cancellable to false when creating a new buffer.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>