Add check for Microsoft Windows Subsystem for Linux
The MSG_TRUNC socket flag when used on Windows Subsystem for Linux (WSL)
seems to result in a returned EINVAL error.  This commit adds a _MS_WSL
preprocessor macro to select the code block that does not require
this flag.  The macro is set by cmake when "Microsoft" is found within
the system version string.

Signed-off-by: JaredD <jareddpub@gmail.com>