Prevent out-of-bounds access in read_each_attr()/iio_channel_attr_read_all()
read_each_attr()/iio_channel_attr_read_all() read all device or channel
attributes at once. The way this is implemented the server sends back a
packet with all the attribute values. For each value it also stores the
size of the value.

The current implementation blindly trusts this value that is store in the
received buffer and increments its pointer position accordingly. But the
data is untrusted data from the network and could contain random garbage.

If the length stored in the buffer is large enough to move the read pointer
beyond the buffer this will cause undefined behavior.

To prevent this make sure that the attribute length stored in the received
buffer is properly validated and read_each_attr()/iio_channel_attr_read_all()
do not access their receive data buffer beyond the amount of data that has
been received.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>