local: Fix low-speed mode
Commit e6754cb84a8b ("local: Return error when no blocks could be
allocated") made the wrong assumption that the BLOCK_ALLOC_IOCTL will
return the ENOSYS error when high-speed mode is not supported. This is not
the case though and the ioctl returns the EINVAL error instead.

This breaks low-speed support since libiio now assumes that high-speed mode
is always supported and something went wrong when trying to allocate the
high-speed buffer and never tries to fallback to low-speed mode.

Unfortunately it is not possible to distinguish between a EINVAL that means
high-speed mode is not supported and a EINVAL that means high-speed mode is
supported but something went wrong when allocating the high-speed buffer.

To avoid this use the BLOCK_FREE_IOCTL to check whether high-speed mode is
supported or not. If high-speed mode is supported BLOCK_FREE_IOCTL will
never return an error, if it is not supported it will return the EINVAL
error.

Fixes: e6754cb84a8b ("local: Return error when no blocks could be allocated")
Reviewed-by: Richard Tresidder <rtresidd@tresar-electronics.com.au>
Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>