appveyor.yml: Add workaround for 32-bit build
For unknown reasons with the latest version of cmake the 32-bit build picks
up the 64-bit version of the lzma library. For all other libraries it
correctly chooses the 32-bit version.

This causes the build to fail. Add a workaround by manually providing the
path to the correct liblzma file.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>