.travis.yml,CI/travis/after_deploy: trigger other builds after libiio
Let's start a "pipeline", where the libiio build triggers a
`libad9361-iio`, which will trigger other builds as well.

We're getting to a point where the lack of automation is starting to hinder
us a bit and changing things in `libiio` are not visible along the way,
until it gets a bit late.

There's a Python script that each job would run, to check if it's the last
job (in the parallel CI run), and the last job, will trigger the next
build.
There's no better way at the moment to do this, without re-factoring the
entire `.travis.yml` file to work with build-stages (a new feature that's a
bit of work to get working properly with our current setup).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>