build,CI/travis: setup CentOS testing
Adapted from blog-post:
 https://derekweitzel.com/2016/05/03/building-centos-packages-on-travisci/

This essentially uses CentOS docker images to run the libiio build and
create RPM packages, which is a bit closer to the real-world case of RPM
usage.

We're not doing any deployments of these files just yet.
This just builds them, and essentially confirms that we can build and
install on CentOS. A later change-set will add deployment.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>