CI/travis: enable errexit & xtrace behavior in scripts
We want scripts to terminate sooner and show traces of where/what failed.
This is to prevent omitting any intermittent failures later.

It's a somewhat common occurrence in shell scripts that a command fails,
but the script continues move on, and final error code (success) is the one
taken into consideration by Travis CI.

The chances of this happening increase as the scripts grow in size.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>