CI/travis/before_deploy: move `grep` expression in `find`
`grep` will return a non-zero exit code that may get caught when `errexit`
behaviour is enabled, so move the name filtering in the `find` expression.
It's also slightly more efficient (computationally).

If `grep` doesn't output anything, that is perfectly fine in this case, and
we want the script to continue running.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>