CI/travis/inside_centos_docker.sh: hack/patch CPackRPM.cmake for CentOS 7
The version of `cmake` in CentOS 7 is broken for packaging RPMs.
It creates RPMs that conflict with other RPMs, because it adds base-system
directories to the `libiio` RPM (e.g. /usr/sbin, /lib, etc).

This is because the CPackRPM.cmake module has an error in the logic.
To workaround it (until fixed), we're patching it before running the cmake
build. This happens in a docker env, so it should be fine.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>