CI/travis/after_deploy: add job that waits 20 minutes
This job works around one of Travis-CI's short-comings: which is there is
no clear/clean way/rule to say "do this after all jobs have finished" in
the current setup.

There are build stages now in Travis-CI, but re-working our current setup
to that mechanism would take about 2-10 days of work.

The `after_deploy` rule would get run for all builds that deploy, and we
only want it to run a single time to trigger the `libad9361-iio` build.

To over-come this, we're adding a separate job that runs first, waits for
20 minutes, and then triggers a `libad9361-iio` build.

It's not ideal, but it should work to have some sort of pipeline for
testing propagation of changes.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>