CI/travis/lib.sh: add utility library & check deploy condition here
This introduces a `lib.sh` that contains all the boiler-plate logic for the
libiio build, and other builds along the pipeline.

For starters, the function is_deployable_travis_ci_run() checks if this is
a Coverity run and that there are keys env vars for deployment. This makes
sure that forks don't try to deploy anything.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>