CI/travis: split CentOS docker build logic into it's own build
This allows the scripts to be run in a similar fashion to Ubuntu, i.e. in a
docker-less environment. This is great if you want to test in a VM, and it
gets us closer to having a unique "inside_docker.sh" script that can handle
just docker-ization.
Though, in order to get there, CentOS has to get better, so that we avoid
having having the CentOS 7 hack.

Note that the test of installing locally the deb,rpm artifacts are
removing, because the new pipeline will make sure that they get propagated
into the next builds, and they will be better tested in those contexts.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>