CI/travis/inside_docker.sh: export TRAVIS_BUILD_DIR var
This offers more robustness when using the docker scripts.
Some scripts in the CI/travis folder may depend (or want to depend) on
this variable during build, so we provide it.
We could typically get away without using this, but it's good to provide
the option of having it (inside dockers).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>