Revert "sort: when sorting iio_channels, if the index is the same, use ID"
This commit broke interop with previous versions of libiio, in certain scenarios.
The interoperability issue is only a problem in case all channels use the same index,
which is typically only the case with the bit mask channels used for logic analyzers.
This means v0.16 can’t interop with previous versions,
since channels are now sorted differently on the local and remote backend.

This reverts commit 67a994c973e68ddcf48c64ebaa0001c69966f6e8.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>