CI/travis/lib.sh: fix triggering multiple builds
Travis CI has multiple job states to account for (queued, received, etc).
In some conditions, the number of created + started == 1, which typically
happens during multiple Travis-CI builds running.

It's also hard to say "job[state] not in [list states]" since there are so
many job states to account for. So we try to account for the ones we could
identify so far.

Travis' API docs are not that great either. They claims it's
self-documenting. Oh well...
Links:
  https://docs.travis-ci.com/user/developer/#api-v3
  https://developer.travis-ci.com/resource/jobs#Jobs

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>