local: Improve support for channels with extended names
On Chrome OS devices, there is a driver named 'cros_ec_sensors_ring',
which defines a channel of type IIO_ACCEL with an extended name of "flag".

When trying to parse the 'en' attribute of this channel, libiio parses it
as a 'flag_en' attribute of the 'accel' channel instead, in turn causing
the context creation to fail due to an unrecognized attribute name.

This patch works around this issue by locating the last underscore as
the channel name marker, instead of the first. This, in turn, yields a
correct channel name of 'accel_flag' and a valid attribute name 'en'.

To the best of my understanding, this should not change the behavior of
libiio for channels that do not have an extended name.

Signed-off-by: Enrico Granata <granata.enrico@gmail.com>