ci,lib: fix typo; correct is `brew list --versions`
The issue only shows up when using the `brew_install_if_not_exists`
function, because that one checks if the package is installed first, and
only then installs it.

`brew ls --version` also seems to be valid, but it looks like it prints
`brew Version` and exits successfully.

For this check, we only need `brew ls`, but using `brew list --versions`
also prints in the log the current installed versions, which is useful when
debugging.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>