DNSSD: Increase DNS Service Discovery robustness
This changes the implementation of avahi and Bonjour to be
more robust, and share more code.

It also now finds all DNS SD hosts advertised on the network, and
builds up a linked list to keep track of them. When you add "ip:" as the
uri, it will connect to the first (random) one it finds. The previous
implementation connected to the last (random) one if found, so I don't
think the overall functionality will be changed in any way.

Signed-off-by: Robin Getz <robin.getz@analog.com>
Signed-off-by: Matej Kenda <matejken@gmail.com>