iio_readdev: Skip output channels
Calling iio_readdev without any channels specified will enable all channels
of the device. Regardless whether the channel is an input or output
channel. Change this so that output channels are skipped.

Similarly ignore explicitly specified channels if they are an output
channel.

The main advantage of this change is that iio_readdev will now exit with an
error when it is accidentally called on an IIO output device, since no
input channels will be found.

But it also future-proofs the application for a time when there is support
for devices which have both input and output channels.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>