add default formatting templates for clang-format and cmake-format
per discussion in #397, The clang-format is the linux kernel code formatting
template with a few minor tweaks:

indent continuation more:

-ContinuationIndentWidth: 8
+ContinuationIndentWidth: 16

remove the kernel's ForEachMacros

Sort the include paths, like Lars suggested in #411

-SortIncludes: false
+SortIncludes: true

In this order:

-#IncludeBlocks: Preserve # Unknown to clang-format-5.0
+IncludeBlocks: Regroup
 IncludeCategories:
-  - Regex: '.*'
-    Priority: 1
-IncludeIsMainRegex: '(Test)?$'
+  - Regex:           '^<.*\.h>'
+    Priority:        1
+  - Regex:           '^<.*'
+    Priority:        2
+  - Regex:           '.*'
+    Priority:        3
+IncludeIsMainRegex: 'xxx'

The cmake-format.py file is from the doc, with tabs set to 4 spaces
https://cmake-format.readthedocs.io/en/latest/configuration.html

The cmake-format.py has been configured to not touch the comments
because some formatting isn't done right. For example, a comment
containing a function and its parameters got formatted and lots
of spaces were introduces making everything unaligned. The comments
format shouldn't be that strict.

This is something for discussion, This is expected to have
clang-format version 7.0.1-8 (tags/RELEASE_701/final) or higher.

Signed-off-by: Robin Getz <robin.getz@analog.com>