ci,build: split trigger-next-build into it's own stage
This change creates a new stage for triggering the next build in the
pipeline and removes the old python script that checks if this is the last
job in the build.

At the current time, I am not sure if the triggering will happen if the
previous stage succeeded or not. But even so, it should be ok, since this
shouldn't occur too often, and would only happen on master and release
branches (which should be stable; and if they aren't this should show).

Also, one main advantage of this approach is that we reduce access to
Github's API which has a rate-limit.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>