ci,build: export DEBIAN_FRONTEND=noninteractive when installing apt packages
On Ubuntu 20.04 it seems that we're getting this during run-time:
```
debconf: (Dialog frontend will not work on a dumb terminal, an emacs shell buffer, or without a controlling terminal.)
debconf: falling back to frontend: Readline
Configuring tzdata
------------------

Please select the geographic area in which you live. Subsequent configuration
questions will narrow this down by presenting a list of cities, representing
the time zones in which they are located.

  1. Africa      4. Australia  7. Atlantic  10. Pacific  13. Etc
  2. America     5. Arctic     8. Europe    11. SystemV
  3. Antarctica  6. Asia       9. Indian    12. US

Geographic area:
```

It may be that Ubuntu upstream may fix this, but it may be an idea to have
DEBIAN_FRONTEND=noninteractive present in the future as well, to prevent
other packages from popping interactive shells.
Hopefully nothing in the interactive shell is important, but configuring
stuff in docker-builds is tricky anyway, since it requires per-package
knowledge [which is bit of work].

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>